Rename 'slave' to 'replica'
Remove all references to the terms 'slave' and 'master, except in code
that is specific to MySQL. Use the terms 'replica' and 'replication
source' instead.

Add a db migration to rename the instances.slave_of_id column to
replica_of_id.

Update apischema.py to reflect support for replicas.

Depends-On: Id22d18a84a4ac104ff8af09959e8eb2ae0102e97
Related-Bug: 1360310
Change-Id: I68f6c5f2fdb210afcbad504df4cf0588400fc42a