Added regression test for config with long value
This bug is fixed by changes from
https://review.openstack.org/#/c/79850

The bug was originally caused by ConfigurationsController._find_type()
only returning an int as the type even for long values.  With the
changes from the above review _find_type() now returns valid types of
int and long.

Added a regression test to validate the scenario described in the bug.

Change-Id: I6df4c919ccccad7481197b8feddaf6f27a5e4e42
Closes-Bug: #1308665