Added support for incremental backup replication
Previously when creating a replica, a full backup was taken.
This could become very time consuming if the dataset is very
large.

This patch changes the default behaviour to first see if a backup
exists.  If one does, it will be used to create an incremental
backup.  This incremental backup will then be used to create the
replica.

A new method has been added to backup/models
(Backup.get_last_completed) which returns the last backup that was
completed for the specified instance.

Change-Id: I2a939bd1a2b7e190502295e6fce08ea0e925472e
Closes-Bug: #DBAAS-190