Implement Oracle 12c proxy guest agent
Oracle 12c is currently not supported in Trove, and we need a
way to expose Oracle Dbaas functionality through Trove,
meaning we need a way to manage (e.g. create and drop)
Oracle 12c databases through Trove.

The requirement is not to host the actual Oracle database
service in the trove guest instance itself.  Instead, the
guest instance will act as a proxy between Trove and a third
party Oracle server which actually host the databases.

This commit contains implementation of this requirement.

Change-Id: If31763436f12e098f4bb126daecead3cc886c15e
Implements: blueprint oracle-proxy