Add support for starting multiple replicas
In the replication v1 API, if you wanted to start multiple replicas
you would have to run multiple 'trove create' calls.  This patchset
adds a 'replica_count' value to the API, to allow multiple replicas
to be started at the same time.

The instances are brought up sequentially (in order to leverage the
existing code base), however once the instances are available, the
trove setup happens in parallel.

Change-Id: I31d522f1bd757b4ee233746709315d12a785d6a5
Closes-Bug: #DBAAS-191
Implements: blueprint replication-v2