Add templates for MariaDB
MariaDB is currently in this limbo state between full datastore and
one that is completely compatible with Mysql. Since it isn't actually
completely compatible, the latest GTID changes for Mysql exposed an
issue where MariaDB ends up with the wrong template. For now, the
solution is to create downstream templates for MariaDB.

Change-Id: I0c1176b8e10c4b12eefa51827a2a36ec621bcc90
Closes-bug: dbaas-373