promote-to-replica-source not reversible
It was not possible to execute promote-to-replica-source on a
replication set which had previously had promote-to-replica-source
executed upon it.  This was because RESET MASTER was being
executed which messed up the transaction numbers.

This change removes the RESET MASTER and causes logging to be
turned on on all slave sites in addition to the master.

Change-Id: I32ae2a5a281701c2f6c4f4d3fbb8de5709904fb3
Closes-bug: dbaas-389