RHEL support for PostgreSQL
Provides RHEL support for PgSQL and makes it easy to extend further
to support other linux variants.

Currently, config.template files for the various datastores are
rendered by the task manager and sent to the guest agent. For
databases such as PostgreSQL that may have OS-specific config
entries, the most appropriate place to render portions of the
config template is on the guest itself.

The template generated by the task manager will therefore leave
any guest-specific elements intact so that they can be rendered
in a second pass.

Once the pgsql guest receives the template, it replaces the OS
specific entries. The same logic is used to 'render' previously
hardcoded constants for things like the location of the
postgresql.conf and pg_hba.conf files.

Closes-bug: DBAAS-769

Conflicts:
	trove/guestagent/datastore/experimental/postgresql/service/config.py
	trove/tests/unittests/common/test_template.py

Change-Id: If9284f91e5cffa14f607f2bf6c50080bbb0abaf0