Fix Cassandra to close connections properly
Cassandra guestagent manager currently does not close
connections when the connection object goes out of scope.
A new connection is created every time the database
status is queried. The manager eventually exhausts the OS
limit on open files.

This patch set fixes a wrong condition check that causes
the issue with closing of active connections.

Change-Id: Ib7b7ba953eb46d39edc8aefba0af647b7a75fefd
Closes-Bug: DBAAS-621