Require password on Cassandra user rename
Cassandra does not have a native way for renaming users.
We implement it by creating a new user, transferring permissions and
dropping the old one (which also removes its existing permissions).

This method, however, requires the user provide a password
(which may be same as the original one).

This patch set adds a check and an explanatory message for the case
when the user provides new name only.

Change-Id: I976b6a57a57384585ffe76afeb593f9011d8bf2a
Closes-Bug: DBAAS-661
Related-Bug: DBAAS-767