Reset 'cluster_name' of a restored Cassandra instance
Instances are created as single-node clusters. A restored instance
should therefore belong to its own cluster as well.
This patch set adds a procedure to change the 'cluster_name'
property of an exesting instance.
The name is updated in both the system keyspace table and
configuration file at the same time.

Change-Id: I187fadc91a9597a8b776366a4803fbd66bd71d78
Closes-Bug: DBAAS-671
Related-Bug: DBAAS-766