Decrease replication slave retry wait time
If the mysql replication slave fails to connect to the master on
its first attempt, it will wait 60 seconds to retry (60 seconds is
the mysql default). This is too long as the guest only waits 60
seconds to determine if the slave is running. It doesn't make sense
to increase the timeout as decreasing the connect retry time will
make the slave get into a the correct state faster. This change
decreases the retry time to 15 seconds which should be more than
enough time to possibly withstand 2 or 3 failures and still get
connected.

Change-Id: Ib53daca645ed05f26258ca6b309ab11b0384edd9
Closes-bug: 1462520