Add support for root-disable
Added the route action for disabling the root user in the extensions.

Modified the resource extension to allow the generation of a DELETE
route on the resource itself.

Implemented root-disable on the mysql guest.  Added not implemented
error messages for all other datastores.

Also needed to add a get_root_user guest api to retrieve the specific
root user for a particular datastore.  Needed this api to generically
find the correct root user for any datastore.  Implemented the
get_root_user for mysql and added not implemented error messages for
all other datastores.

Change-Id: I52519b86c47694c554b624d1d2fbe7a001af55fc
Partial-Bug: DBAAS-685