PostgreSQL guest agent can't remove temp file
The postgresql guest agent uses the pgsql COPY TO command to extract
and cache db resultsets in temp files. These are used by the user
management commands like user-create and user-list.

Since the sticky bit is typically enabled on /tmp, attempts to
remove the file with the user running the guest agent service fail
with an 'Operation not permitted' error.

Closes-bug: #1464389
Change-Id: I15452e7a2bc9cb6823ce2154f8c03c6c6a0b988b