Validate user names per datastore in the API layer
User names are only validated against the MySQL user model in the API
layer. This isn't useful for some datastores, e.g. Mongo. This can lead
to an instance create failing after an extended period of time with no
information available to the administrator.

Fix is to validate based on the datastore type of the instance.

Change-Id: I417e6d5c8f35a6ff7df3489eba3d08d9f68abb05
Closes-bug: dbaas-826