MongoDB create_admin_user not authorized
During cluster creation the taskmanager attempts to run
create_admin_user on the query router, but it fails as the connection is
not authorized if security/authorization is enabled. This causes the
cluster creation to fail.

The fix is to pass the localhost as the host for the client, then reset
it to the host address after.

Change-Id: I95083d0981231bfd19e9646446c25d0b3f112b12
Closes-bug: #1485232