Fix parent['id'] KeyErrror for replication backup
Some backup code paths (e.g. replication) does not pass in the new
parent['id'] parameter, which was introduced to support the new
"--parent last" switch for trove backup-create.

This commit added code that checks for existence of parent['id']
before accessing it's value, which prevents the incremental backup
for replication from crashing.

Change-Id: I441ffd7eaec44f1990fd7b1b39c086081df7a1fe
Closes-Bug: dbaas-839