Determine Mongodb's PID file location dynamically
Mongodb by default does not start as a service.
In order to start it as a service we need to define a PID file
so that the OS's service manager can find its process.

The location is currently hardcoded to '/var/run/mongodb.pid' -
path used by Ubuntu.
RHEL-based systems however store it in '/var/run/mongodb/mongodb.pid'.
We need to pick the correct location in the GA.

Change-Id: I93055d726b5e678e5a5bfe033fbcc3392eb6b70e