Fix ORA-00922 error on Oracle CREATE USER statement
A previous commit removed the wrapping double quotes around the password
parameter in the CREATE USER statement. The change caused any passwords,
including the system generated admin user password, that starts with a
number to fail the CREATE USER statement with a ORA-00922: missing or
invalid option DatabaseError.

This commit resolve the issue by restoring the wrapping double quotes
around the password.

Change-Id: Ie12560df266f147679fb18d56c99b3fd7f67f351
Closes-bug: dbaas-856