Fix resize-volume error for Oracle11g on RHEL
e2fsck needs to be run with the -p option before the resize2fs command.
The current code runs e2fsck with the -n option for RHEL, causing the
resize-volume for Oracle11g on RHEL to fail.

e2fsck also requires the volume to be unmounted, some distro (e.g. RHEL7)
automounts the volume and therefore causing e2fsck to fail.

This commit contain changes that address the above issues.

Change-Id: I3c73fe8f61068a70025a4b4ac9342d12891883ca
Closes-bug: dbaas-852