Refactor the datastore manager classes (1+2)
There is a large amount of boiler-plate code in each datastore manager.
As more managers are added, the time involved in maintaining all this
code wil continue to grow. To alleviate this, a base manager class
has been added where all common code can reside.

The initial refactoring just moved some of the obvious code (such as
rpc_ping and update_status) into the base class, along with defining
properties that can be used to further abstract functionality
going forward.

The issue of having instances move in-and-out of ACTIVE state has
also been fixed by adding a flag file that is written by the base
class once prepare has finished successfully.

This combined changeset moves most of the rest of the boiler-plate code
into the base class.

This includes more filesystem methods, the 'unsupported'
methods and a configuration method.
It also consolidates the replication strategy.

The prepare_completed flag was also prefixed with a
double underscore to make it difficult for descendant
classes to change it by accident.

Log messages were also enhanced.

Change-Id: I80944c58e10d7de5719201afc8b04d0cc9a0209f