Fix Oracle 11g backup and resize-volume errors
1) Changed from using trove.openstack.common.log to oslo_log.

2) Status of Trove instance should be BUILDING instead of SHUTDOWN
when building of Oracle database instance is in progress.

3) Changed the backupagent unit test to address failures caused by
the new Oracle backup/restore feature.

4) Fix resize-volume error for Oracle11g on RHEL:

e2fsck needs to be run with the -p option before the resize2fs command.
The current code runs e2fsck with the -n option for RHEL, causing the
resize-volume for Oracle11g on RHEL to fail.

e2fsck also requires the volume to be unmounted, some distro (e.g. RHEL7)
automounts the volume and therefore causing e2fsck to fail.

5) Fix ORA-00922 error on Oracle CREATE USER statement

A previous commit removed the wrapping double quotes around the password
parameter in the CREATE USER statement. The change caused any passwords,
including the system generated admin user password, that starts with a
number to fail the CREATE USER statement with a ORA-00922: missing or
invalid option DatabaseError.

Change-Id: Ibf8c8bce49d2c25ce690ed77f186b759f2a46491
Closes-bug: dbaas-836, dbaas-856, dbaas-852, dbaas-945