Add unspecified options to backup unittests
The backup runner unittests have been intermittently failing
because some of the backup commands being returned don't include
gzip in the pipeline.

The problem is that the tests all use the same instance of the
BackupRunner and the is_zipped instance variable is being
altered by the various tests. Each test that is expecting
compression and/or encryption needs to set these variables
explicitly. If not, depending on the timing and ordering of
the tests, the variables may not be set as the test expects.

Change-Id: I74431aefe88e1aea15ca5160a6cb597b59d19fbd
Closes-bug: 1500199