Fix issue where log file could not be read
log-publish would complain sometimes that the file could not be read.
The directory has been changed to have the correct permissions to
allow the file to be accessed properly.

Also changed the guest definition to pull the log_dir and log_file
from the guestagent conf settings.
(/var/log/trove/trove-guestagent.log is the default).

Change-Id: I0a71213f543e93a3c1b4e2ae49ebafb6dc838631