Changed postgresql to load changes dynamically
The order of applying changes to the configuration were such that
postgresql required a restart (the order doesn't matter for mysql
due to the nature of how config changes are applied - with
postgresql they must be made first in the config file as the server
is told to load this).

The order has been reversed so that postgres loads the correct
values.

Change-Id: I8ea80e002988eb28d2a48fbfa417fe2579573c15