Improved performance of publish by caching states
The code was repeatedly running chmod on the log files.  This turns
out to be a very expensive call (~ .5s) that was making the command
seem overly sluggish.

Caching was done on the enabled state and whether the file has
been marked as readable.

The 'status' field was updated too, since there is now an 'enabled'
flag.

Change-Id: Ibef5022b0cdbffbb98aa6df21241b636a47df9c6