Expand log commands to allow break-out of fns
The original log-publish command was difficult to explain,
so this change req breaks it down into 4 parts:

log-enable
log-disable
log-publish
log-discard

The first two enable/disable the actual logging done by the
database.  The last two publish the log to Swift and
delete the Swift container that holds the log.

The API was changed to be guest_log_action and 4 flags are now
passed to it to correspond with the commands above.

Change-Id: I9bb8a13cc9cdf23e68d8c3cfad084f1d8300d697