Fixes slow_query log enable issue for MySQL
Enabling a slow_query log on MySQL generated an error from the
MySQL client due to the fact that the slow_query variable
is a float (and was being treated as a string).
Also enabling a log twice caused an empty result payload to
be returned (generating an error in the client).

Both theses issues have been fixed.

Closes-Bug: #DBAAS-1039
Change-Id: Iedc4394bbf5bb46779d51c3ce6e9d7f7e6dee4b9