Make sure Redis is restarted after conf changes
In prepare, in order to make config changes on the server, it must
first be restarted so that the mapped 'SET' command is recognized.

This happens with a restore; otherwise it must be done explicitly.

Not sure why this isn't in the repo - it was pushed upstream, and
the changes made it there.  For some reason only half made it DS ...

See: https://review.openstack.org/#/c/226075

Change-Id: I2cc92edd9abd35c84d3094fab0bb07ea68d1f373
Closes-Bug: DBaaS-1012