Fix for oracle_ra to parse config_contents properly
In _do_load_oracle_config() the config_contents now come in as a
dict so there is no need to write the temp config file and parse
the contents of the temp file for the config parameters.

Use the dictionary directly now and remove the code to write the
temporary file.

Change-Id: Ibadf321c2bd08df97713b6882799cf42bd4dbb05
Closes-Bug: DBAAS-931