Remove non-working logic for 3+ node configs
There was logic in place to attempt to handle the situation where
a third slave needed to switch masters. Not only was this
insufficient due to how postgres handles timeline jumping, it also
caused successive switches between master/slave to fail.

This should enable a working master-slave setup with the ability
to switch back and forth indefinitely. A follow up fix will
deal with the more complex scenarios

Change-Id: I545f40c5aafda2e43d2398a278376ce9f9c9cff5
Fixes-bug: DBAAS-1084