Fix race condition in replication backup delete
In the int-tests, the replication implementation verifies that the
backup used to create the replica is removed after the slave goes
active. This test runs right after the slave create, so on occasion
the backup still exists when the test runs, since the taskmanager
doesn't delete it until it also receives word that the instance is
up and running.

The test was placed lower in the test run (right after the slave
delete), which should alleviate the problem.  Code was also put in
place to take into account the fact that a backup may already exist
(as is done in the backup/restore tests).

The replication scenario test was also modified to include this
check at the end of its run.

Closes-Bug: #1518477
Change-Id: If815f496d6bf9111dd1f04bf02f61207a2317b2b