Oracle configuration group support
Allow the Oracle datastore to be configured via configuration groups.
key/value pairs passed in through a group map to Oracle initialization
parameters. These will be stored in an Oracle PFILE, which will be used
to generate the default SPFILE. This allows the administrator to change
Oracle settings. Settings that would interfere with Trove are not
supported.

This change enables the codepaths for overrides on the guest. PFILE and
SPFILE management is done by the guest. PFILE versioning is handled by
the configuration manager. A list of supported parameters is included in
the Oracle templates directory.

Change-Id: I2d61d77ade244adcd2a61d110a6f94e83f4aaa41
Implements: blueprint oracle-config-groups