Workaround for resize-vol issue on RH
HACK ALERT...the resize-volume issue on Centos/RHEL can be
worked around by putting a small sleep in the task manager
resize method. This adds a 2 second sleep.

The real solution appears to be to implement the use of
systemctl mask <mount-point> but that seems like a pretty
involved platform specific fix that involves more thought.

Change-Id: I8b2c3422796a31c0daf6fd87fb200d0e9bd21055
Closes-bug: DBAAS-1037