Get Oracle database name from oratab
The database name is currently assumed to be the same as the instance.
In a restored instance, this may not be true, and therefore some guest
commands are broken. The fix is to get the database name from Oracle's
list - the oratab file. This is already done for list_databases, so just
call that.

Change-Id: Ia87ae5e178f35cf1fff99bb893db752ae063f978
Closes-bug: dbaas-1118