Fix to_gb & to_mb conversion function rounding issue
The to_gb() and to_mb() conversion functions rounds the result to 2
decimal places. Results less then 0.01 are being rounded down to 0,
causing the redis backup test to fail.

This commit ensures the functions only return zero if the input is
zero, and a value of at least 0.01 will be returned for any value
bigger than zero.

Change-Id: I9e520e3611984d6a21217cced1b08b311aa27205
Closes-Bug: #1531257