Fix PXC to work with v5.6
The upstream implmentation of Percona clusters only supported
PXC 5.5. To make the implementation work with v5.6, some changes
needed to be made to the base Mysql manager. The biggest change
is that we can no longer store the os_admin client credentials
in the server my.cnf. This commit moves those credentials to
a .my.cnf file stored in the trove homedir. This will be read
by the guest when attempting sqlchemy connections to the local
mysql database and will also be picked up automatically by other
mysql clients (such as mysqladmin).

Change-Id: Ie98ea4ff13423e17f17169eaa36d8946f58ff7b9
Closes-bug: dbaas-1017