Sometimes prepare messages are delayed
This bug addresses one issue which manifests itself as a delayed
deliver of prepare messages on the guest. A full description of the
change is found in the bug report along with a set of sample programs
to recreate the problem and test the fix.

I'm marking this as 'partial-bug' only because I'm not sure this is
the only issue that can cause this. It is just that this one is
certainly the one that has been plaguing me on my dev environment.

Change-Id: I065fb7c7f674ce587d4a480ac2f3462826ba2604
Partial-Bug: #1511623