Fix backup of mysql variants on RHEL
The switch to XtraBackup 2.3 was causing an issue on RHEL/Centos
related to is using the "mysql" user instead of "trove". The fix on
Ubuntu was to move the os_admin credentials to ~trove/.my.cnf. While
this is a better place to write the credentials anyway (i.e. they
shouldn't be stored in server my.cnf) this doesn't solve the whole
issue on RHEL. This commit changes the XB backup strategy to pass
the user/password in on the innobackupex command line.

Also, it was noticed that the "socket" option wasn't being specified
in the config.template. This is causing some client connections,
such as XB to fail connect because it can't locate the socket.
Forcing the server/client to write/read the socket from a known
location fixes this.

Change-Id: Iea941ce60179ef4dc5c403c2fc374cc8eb7d1617
Closes-bug: dbaas-1264