Handle Oracle configuration parameter errors
Due to a bug in oslo_messaging some exceptions raised by the GA are not
reported back to the api/taskmanager. In the case of the API this can
cause a hang that ends with a timeout.

Oracle configuration parameters are not vetted by Trove and may be
invalid. This sometimes results in the above scenario during
configuration-attach and -patch.

The fix in this case is to handle the exceptions raised by cx_Oracle
relating to invalid configuration parameters. Following the lead of the
MySQL GA code, each parameter is applied one at a time and invalid
values are logged and skipped.

Closes-bug: DBAAS-1325
Change-Id: I92740b976a06d0bb7d2ad1073111f72003f8f9ac