Fix Oracle replication failure after root-disable
Replica creation fails on the slave node with this error after
root-disable the master:
ORA-01017: invalid username/password; logon denied

The sys password updated by root-disable was not updated to the
cnf file and as a result, the new slave tries to connect to the
master with the old sys password when cloning data, leading to
the error. This commit ensures the sys password in the cnf file
is updated after root-disable.

Change-Id: I186ed19b38a8fcb56547dde7fbfa88e97afe895e
Closes-Bug: DBAAS-1445