Fix Oracle replication failure after root-enable
Replica creation fails on the slave node with this error after
root-enable the master:
ORA-01017: invalid username/password; logon denied

The sys password updated by root-enable was not updated to the cnf
file and as a result, the new slave tries to connect to the master
with the old sys password when cloning data, leading to the error.
This commit ensures the sys password in the cnf file is updated
after root-enable.

Change-Id: I497aab8a39940b7a7ddf84ba09d20920d297aaf7
Closes-Bug: DBAAS-1445