Fix inappropriate Oracle replication SHUTDOWN status
When Oracle promote-to-replica-source (i.e. switchover) finishes, the
status of the new master would sometimes become SHUTDOWN before turning
into ACTIVE. The SHUTDOWN status is the result of Oracle service reset
during the switchover process. This status shouldn't have been casted to
the client and visible to the end user.

This commit address the issue by updating the instance's status to the
actual value as soon as switchover completes.

Change-Id: Iaab627879f8fbc17fc67df776aaf3725a4829de6
Closes-Bug: DBAAS-1452