Fix MariaDB config groups
MariaDB historically leveraged the mysql manager for guest
support including the configuration groups implementation.
With MariaDB now having it's own manager class that inherits
from Mysql, it needs to have validation_rules and a
ConfigParser setup.

This commit adds those items and turns back on the configuration
scenario tests (which were run with MariaDB to test this change)

Change-Id: Iaae1bb3916f0d05f7d1566dbeb39903763447657
Closes-bug: 1532256