Add log retrieval to Cassandra
Add system.log retrieval.

The Cassandra's system/general log is located in the cassandra
log directory and is by default called system.log.

Cassandra 2.1 and higher uses 'SLF4J' with 'logback' backend.

Logging properties can be configured at runtime via the nodetool utility
(setlogginglevel) and/or persisted in 'logback.xml' configuration file.

Logging can be enabled on per-module/class basis with configurable
verbosity levels.

The logging in Trove is configured globally (on the root level)
- i.e. for all application modules.

A new configuration property 'system_log_level' was added to control the
log verbosity. It allows the operator define the log level that will be
set when the logging is enabled on an instance.
All logging will be turned off when disabled.

Note: The log level is not changed during prepare. The instances will be
provisioned with the default logging setting (INFO level).

Note: An XML codec was added to allow parsing XML files into/from Python
dicts.

Closes-Bug: 1550557

Conflicts:
	requirements.txt
	trove/common/stream_codecs.py
	trove/guestagent/datastore/experimental/cassandra/manager.py
	trove/tests/scenario/runners/guest_log_runners.py

Change-Id: I1532cd2ff24fc793e8e49eee81cda47c7b6e2ae2