Add log retrieval to Cassandra
Add system.log retrieval.

The Cassandra's system/general log is located in the cassandra
log directory and is by default called system.log.

Cassandra 2.1 and higher uses 'SLF4J' with 'logback' backend.

Logging properties can be configured at runtime via the nodetool utility
(setlogginglevel) and/or persisted in 'logback.xml' configuration file.

Logging can be enabled on per-module/class basis with configurable
verbosity levels.

The logging in Trove is configured globally (on the root level)
- i.e. for all application modules.

A new configuration property 'system_log_level' was added to control the
log verbosity. It allows the operator define the log level that will be
set when the logging is enabled on an instance.
All logging will be turned off when disabled.

Note: The log level is not changed during prepare. The instances will be
provisioned with the default logging setting (INFO level).

Note: An XML codec was added to allow parsing XML files into/from Python
dicts.

Change-Id: Ia57587bf557771ea6d7f00e2fe9d674789b98559
Depends-On: I4a088b7a541c35a6c32d6985727bad65ff491815
Depends-On: I254b81b45e44aeda3049865bee76cd9075312761
Closes-Bug: 1550557