DSE guest-agent fixes
The current implementation of DSE Manager calls the super
init which creates App and Admin objects and passes
the manager's name to the base Manager.
The app/admin built during initializations are the community
Cassandra implementations not necessarily compatible with
DSE.
The manager name passed to the base is also
the name of the community manager ('cassandra').
This may result in unexpected behavior in the future
(loading of wrong configs).

The patch includes following fixes:

 - Pass the proper manager name to the base Manager.
 - Build the App and Admin object via overridable interface.

Depends-On: Id448a494600a3033cc06ebbb45cd13d68e233262
Depends-On: Ie10c5cc4d5390013d111aef27991a747ac2ff795

Change-Id: I3a3eb872bf005adc15e34ee3cd0c7fd6f6449f52