Cap test_log_generator_user_by_row to 100
For some datastores (for example postgreSQL), the guest-log scenario
test test_log_generator_user_by_row can take an inordinate amount of
time, since a lot of data is written into the user log once it is
enabled. This test should be capped so that an extremely large log file
won't cause the tests to time out.

Change-Id: Ie10c5cc4d5390013d111aef27991a747ac2ff795
Closes-Bug: #1559214