Fix Oracle detach replica orphaned log dest and other cleanups
This patch reapplies commit 305b8c3cfeee401f, which was
previously reverted. The following is the detail of the original
commit:

Currently, detaching an Oracle slave only turn off the log apply
process on the slave, logs are continuously being shipped to the
detached slave and accumulates there because configuration of
the slave being detached still exists on the master and the
other active slave nodes.

This patch changes the way log shipping destinations are being
maintained. The master node will be the only node that contains
the latest list of slave destinations. This list will be copied
to the new slave in switchover scenarios. When a slave is being
detached, it will be removed from the log shipping destination
list on the master.

Changes has also been made to eliminate the enable_as_master
"for_failover" parameter. Also deleted the enable_as_master_s2
function since it is no longer needed after the removal of
"for_failover" parameter.

Change-Id: Id88f19c0a3d354cbb30020b644e1b3e6409a60bc
Closes-Bug: bug/dbaas-1369