Fix MariaDB clusters failing with TypeError
This would apply to any type of cluster that uses the galera strategy
while setting the nics on a create call. When we called cast to set()
the object was a list of lists. The set method can not has a list so
this was causesing a unhashable error.

The change is to make the instance_nics a list of strings (what we
originaly expected) to resolve this issue.

Change-Id: I77cc008243b4c9bbfceeb2a67dddbca856b7a37c
Closes-Bug: #DBaaS-1772