Refactor scenario tests to facilitate multi-group
The scenario tests need to have the group classes broken
up to allow interleaving of the different parts of the tests
so that the gate can finish in a reasonable amount of time.

The runners, however, need to be the same for each main group,
thus allowing state to be preserved between groups of the same
feature.  The creation of runners was refactored to allow for this.

The module tests were also modified to use the new paradym, which
allowed the removal of the global varaibles in the runner.

Dependencies on tests.api were also removed.

Partial-Bug: #1571092

Conflicts:
	trove/tests/int_tests.py
	trove/tests/scenario/groups/root_actions_group.py
	trove/tests/scenario/runners/instance_create_runners.py
	trove/tests/scenario/runners/test_runners.py

Change-Id: I8bf4f50e97417ddbf7f3951dd2ce481a400d53ac