Notification exceptions not sent back correctly
With the new notification feature added recently, exceptions
that occur on the guest are sent back to the conductor.
Due to a mis-match in the code however, the wrong call is being
issued. The conductor api makes a call to notify_exception
instead of notify_exc_info.

This has been fixed.

Change-Id: Ib2e04bb903c92b4d570913aba2e985db63c45096
Closes-Bug: #1577848