Catch proper exception on root-enable
The mitaka install guide recommends installing python-pymysql on
Ubuntu and python2-PyMySQL on RHEL/centos.

The pip library that gets installed due to the packages above are
PyMySQL==0.7.2 in Ubuntu and PyMySQL==0.6.7 in RHEL/centos.

The root-enable call attempts to create a user and catches a
specific exception before continuing on.  Since not all the
different environments are updated to use the new library
we will catch both the old and new exception.

Changed code to catch the InternalError as well.

Change-Id: I4a1315cd93e8debde760df46021fc909fca170eb
Closes-Bug: DBAAS-1500