Data volume not always remounted after upgrade
Some cinder backends return a different form for the "device"
parameter than the code was expecting. It expects it to be a valid
file system path to the attached device (e.g. /dev/vdb). Some
backends return just the "vdb". Code was adjusted to handle this
and a unit test was added.

Change-Id: I3d749484e81498e99e1676ee2d7a09745b6c1bdd
Closes-bug: dbaas-1784