Persist error messages and display on 'show'
When an error occurs in Trove, it is very difficult
to determine the cause without access to the
server logs. To make these errors available to
the end user, they are now persisted in the database
and can be viewed using the standard 'show' command.

Note: This commit also includes the module db
migrate file, so that the version numbers stay
in sync with 1.8/master/upstream. This means
the two module tables will exist in 1.7 however
there is still no code to support modules.

Change-Id: I4c5c8b7dfc169a4948aa3aa0e9440faac803289d
Implements: blueprint persist-error-message
Closes-Bug: #DBaaS-1099