Allow Mariadb to use defaults for pid/sock files
With the change of MariaDB to use the new mariadb.service file
for startup on Centos/RHEL [1] there are now issues with Trove
setting ths pid and sock files to be written to /var/run in
config.template. It looks like MariaDB defaults these to
reasonable locations (/var/lib/mysql) so let it use those.

[1] https://review.elasticdb.org/2648/

Change-Id: I0b9cd0feee3ff222c27a58514c7afa956a001a11
Closes-bug: dbaas-1702