Use portable queries to update user properties
MySQL guest currently updates user properties by
changing fields in the internal 'mysql' database.

Definition of this system table has changed in MySQL 5.7
and the current code no longer works.

This patch set replaces the UPDATE with more generic
SET PASSWORD and RENAME statements that should work on
all MySQL versions.

Another benefit of this solution is that we no longer need to
transfer privileges on a user name/host change since the RENAME
statement does it for us under the covers.

Change-Id: Id553314e029c637c9f2a8bb3bb6a9d28d63ff510
Closes-Bug: DBAAS-174