Fix MariaDB restore and replication
A previous fix that adjusted the service candidates for
MariaDB 10.1 on Centos/RHEL has caused an issue with restore.
The root causes is that the mysql restore strategy is using
the MysqlApp object. Since MariaDB has its own manager, we need
to change the restore strategy to use the correct App object.

Due to the change to service candidates, the restore and
replication strategies needed to change. Basically, now that
MariaDB is a full manager, it needs to have its "App" class
set properly (this is the restore strategy fix). Replication
broke because the replication strategy was using the Mysql
manager backup/restore strategy. This was fixed by doing a small
re-factor to add properties to the strategy base class to allow
for manager specific overrides.

I ran the MariaDB scenario tests and found that we need
to skip the _validate_binlog_format check during the replication
tests. See [1] for more details.

[1] https://review.openstack.org/#/c/300253/

Change-Id: I0cdb75f750004025b2f4655d2266096da2fda08d
Closes-bug: dbaas-1840