Pass AZ and NIC values to all mongodb instances
The query routers and config servers were not being passed the AZ and
NIC values provided. For each of the 3 instances provided the config
servers will each be given one of the AZs, and the same goes for the
query routers. If additional query routers are provided then the AZ list
will be looped through.

A check that all provided NICs are the same has been added, and the
config servers and query routers will use the NIC value.

Closes-bug: dbaas-1851
Change-Id: I1a5452667e939e41adbdf781084ba5b29763671d