Fix DBCA error on Oracle12c
Linux sets the shared memory /dev/shm to be slightly under 50%
of physical memory by default. However, Oracle use a different
way to calculate the 50% memory of physical memory needed for the
SGA, which ends up to be over the default size of /dev/shm, causing
DBCA to fail.

This commit lowers the memoryPercentage DBCA parameter from 50%
to 40% to avoid overwhelming /dev/shm.

Change-Id: Iac28069737b102fada6dcadd15e34bc35c68256e
Closes-Bug: dbaas-1867