Fix Couchbase password reset
The current code is trying to reuse methods.
Break down the 'update_admin_credentials'
method into two each of which does its own thing.

The 'root_enable' should now always reset the password,
while 'secure' just stores it on the FS and leaves it on
the initialization to pick it up.

'apply_guestagent_configuration' was dissolved into
the Manager code. It never applied any configuration
anyways.

Closes-Bug: DBAAS-1868
Change-Id: I61dbc249cf997a65412382a36a3529f3c553722a