Implement error handling when root enabling an Oracle slave
Oracle replication slave databases are read only. Any passwords
(including the sys password) in such database cannot be changed.
However, trove root-enable allows a user to change the sys password
on such read only slave databases, and actually attempt to change
it behind the scene, which cause errors on the guestagent and sends
an ambigous trove client timeout error message back to the user.

This commit implement checks on the open status of a database before
attempting to change the sys password. If the database is read only,
password update will not be performed and a meaningful error message
will be displayed to the user.

Change-Id: I5e52b14060de00fcd7fbdae25d839d2b74001921
Closes-Bug: dbaas-1425