Fix secure method to work with PXC
When PXC was implemented, the secure method was overridden as the
stock one was causing problems on prepare. This issue seems to
be caused by the order of work done in the call. If the admin
user is set up first and then used to delete the anon users,
everything works for all flavors of MySQL.

Change-Id: I7eb889d916e717d7889682914efaca80a78b56c0