Fix issue with os_admin not being able to connect
With the introduction of the pxc fix and using special
characters for the os_admin password, there was a case where
if the password had a '+' in it, the connection as os_admin
would fail in the new pxc code, since it was urllib quoting the
password.  This has been fixed and the code made to look more
like the US version.

Change-Id: Ibe12d260fe1380da5efae27172d744046cc602e4