Another attempt to fix CB 3.0 backup/restore
Our attempts to have things run as root are just causing
more issues. The root problem appears to be that on first
run of cbbackup/cbrestore, it needs to generate a C++ library
into a cache directory. The "trove" user couldn't do this because
of file permissions (diretory is under /opt/couchbase).

A commit[1] to the packaging repo has caused that directory to
be pre-created with appropriate permissions so we don't
need to run cbbackup/cbrestore as root.

[1] tesora-dbaas-trove-packaging:
    f0380aea5f7c5d37c22d95a8a8c43b770c2b5748

Change-Id: I6ce5752141f1c4ee199710f04eaf3ee6f9e4d970
Parial-bug: dbaas-1931
(cherry picked from commit 2648e1be299a0abaa21d7ce327ee907330af3019)