Compatability changes to TroveContext
In order to facilitate a 1.7 guest running against a 1.8
controller, we need to adjust the context object to handle
with the interface changes introduced by new versions
of oslo.context.

This is for Paypal but should also allow for general 1.7.x
guest against 1.8 controller.

We should plan to remove this code in a future 1.8.x
release.

Change-Id: I1958d11836d802ff70ff7012d0366af9f5ee4f8f